OAK-10147: Many move operations may consume a lot of memory
Combine move tests in a single class and run with parameters for ClassicMove and NeoMove
